Fix entry-points
Imagine a module 'foo' containing files 'index.js' which requires
'./sub.js'. If we want to pack the module and all deps, we want it so
that require('foo') includes 'index.js', not 'sub.js'. However, deppack
currently takes this path:

  loadDeps('foo/index.js')
  readFileAndProcess('foo/index.js')
  loadDeps('foo/sub.js')
  readFileAndProcess('foo/sub.js')
  done('foo/sub.js')

The done() call doesn't actually take any arguments, but the done that
is called is in the scope of the second loadDeps. It seems that when we
arrive at the last module we read in that doesn't have any dependencies,
we use *that* as the entry point name, which seems wrong.

Just use the entry point name, as above, for correct behavior.